renaming clientapp to administration
