+++ NICE.Identity/Administration/.eslintrc
