+++ NICE.Identity/Administration/package.json
