+++ NICE.Identity/Administration/public/manifest.json
