+++ NICE.Identity/Administration/src/App.test.js
