+++ NICE.Identity/Administration/src/Pages/FindUser/index.js
