+++ NICE.Identity/Administration/src/components/FindUserControls/components.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+import { FormGroup } from "reactstrap";
+import { Row } from "react-bootstrap";
+
+export const Container = styled(Row)`
+  min-height: 12.5rem;
+  margin: 1.5rem 0 1.5rem 0;
+  display: flex;
+  padding: 2rem 0;
+`;
+
+export const FormGroupStyled = styled(FormGroup)`
+  margin-left: 1.5rem;
+`;
