+++ NICE.Identity/Administration/src/components/FindUserControls/index.js
