+++ NICE.Identity/Administration/src/components/Layout/index.js
