+++ NICE.Identity/Administration/src/index.js
